move attribute validation methods to node-api engine module
