+++ crates/node-api/src/engine.rs
@@ -1,4 +1,8 @@
-use crate::{PayloadAttributesTrait, PayloadBuilderAttributesTrait};
+use crate::{
+    AttributesValidationError, PayloadAttributesTrait, PayloadBuilderAttributesTrait,
+    PayloadOrAttributes,
+};
+use reth_primitives::{ChainSpec, Hardfork};
 
 /// The types that are used by the engine.
 pub trait EngineTypes: Send + Sync {
@@ -13,4 +17,174 @@ pub trait EngineTypes: Send + Sync {
         + std::fmt::Debug;
 
     // TODO(rjected): payload type
+}
+
+/// Validates the timestamp depending on the version called:
+///
+/// * If V2, this ensure that the payload timestamp is pre-Cancun.
+/// * If V3, this ensures that the payload timestamp is within the Cancun timestamp.
+///
+/// Otherwise, this will return [EngineApiError::UnsupportedFork].
+pub fn validate_payload_timestamp(
+    chain_spec: &ChainSpec,
+    version: EngineApiMessageVersion,
+    timestamp: u64,
+) -> Result<(), AttributesValidationError> {
+    let is_cancun = chain_spec.is_cancun_active_at_timestamp(timestamp);
+    if version == EngineApiMessageVersion::V2 && is_cancun {
+        // From the Engine API spec:
+        //
+        // ### Update the methods of previous forks
+        //
+        // This document defines how Cancun payload should be handled by the [`Shanghai
+        // API`](https://github.com/ethereum/execution-apis/blob/ff43500e653abde45aec0f545564abfb648317af/src/engine/shanghai.md).
+        //
+        // For the following methods:
+        //
+        // - [`engine_forkchoiceUpdatedV2`](https://github.com/ethereum/execution-apis/blob/ff43500e653abde45aec0f545564abfb648317af/src/engine/shanghai.md#engine_forkchoiceupdatedv2)
+        // - [`engine_newPayloadV2`](https://github.com/ethereum/execution-apis/blob/ff43500e653abde45aec0f545564abfb648317af/src/engine/shanghai.md#engine_newpayloadV2)
+        // - [`engine_getPayloadV2`](https://github.com/ethereum/execution-apis/blob/ff43500e653abde45aec0f545564abfb648317af/src/engine/shanghai.md#engine_getpayloadv2)
+        //
+        // a validation **MUST** be added:
+        //
+        // 1. Client software **MUST** return `-38005: Unsupported fork` error if the `timestamp` of
+        //    payload or payloadAttributes greater or equal to the Cancun activation timestamp.
+        return Err(AttributesValidationError::UnsupportedFork)
+    }
+
+    if version == EngineApiMessageVersion::V3 && !is_cancun {
+        // From the Engine API spec:
+        // <https://github.com/ethereum/execution-apis/blob/ff43500e653abde45aec0f545564abfb648317af/src/engine/cancun.md#specification-2>
+        //
+        // 1. Client software **MUST** return `-38005: Unsupported fork` error if the `timestamp` of
+        //    the built payload does not fall within the time frame of the Cancun fork.
+        return Err(AttributesValidationError::UnsupportedFork)
+    }
+    Ok(())
+}
+
+/// Validates the presence of the `withdrawals` field according to the payload timestamp.
+/// After Shanghai, withdrawals field must be [Some].
+/// Before Shanghai, withdrawals field must be [None];
+pub fn validate_withdrawals_presence(
+    chain_spec: &ChainSpec,
+    version: EngineApiMessageVersion,
+    timestamp: u64,
+    has_withdrawals: bool,
+) -> Result<(), AttributesValidationError> {
+    let is_shanghai = chain_spec.fork(Hardfork::Shanghai).active_at_timestamp(timestamp);
+
+    match version {
+        EngineApiMessageVersion::V1 => {
+            if has_withdrawals {
+                return Err(AttributesValidationError::WithdrawalsNotSupportedInV1)
+            }
+            if is_shanghai {
+                return Err(AttributesValidationError::NoWithdrawalsPostShanghai)
+            }
+        }
+        EngineApiMessageVersion::V2 | EngineApiMessageVersion::V3 => {
+            if is_shanghai && !has_withdrawals {
+                return Err(AttributesValidationError::NoWithdrawalsPostShanghai)
+            }
+            if !is_shanghai && has_withdrawals {
+                return Err(AttributesValidationError::HasWithdrawalsPreShanghai)
+            }
+        }
+    };
+
+    Ok(())
+}
+
+/// Validate the presence of the `parentBeaconBlockRoot` field according to the payload
+/// timestamp.
+///
+/// After Cancun, `parentBeaconBlockRoot` field must be [Some].
+/// Before Cancun, `parentBeaconBlockRoot` field must be [None].
+///
+/// If the engine API message version is V1 or V2, and the payload attribute's timestamp is
+/// post-Cancun, then this will return [AttributesValidationError::UnsupportedFork].
+///
+/// If the payload attribute's timestamp is before the Cancun fork and the engine API message
+/// version is V3, then this will return [AttributesValidationError::UnsupportedFork].
+///
+/// If the engine API message version is V3, but the `parentBeaconBlockRoot` is [None], then
+/// this will return [AttributesValidationError::NoParentBeaconBlockRootPostCancun].
+///
+/// This implements the following Engine API spec rules:
+///
+/// 1. Client software **MUST** check that provided set of parameters and their fields strictly
+///    matches the expected one and return `-32602: Invalid params` error if this check fails. Any
+///    field having `null` value **MUST** be considered as not provided.
+///
+/// 2. Client software **MUST** return `-38005: Unsupported fork` error if the `payloadAttributes`
+///    is set and the `payloadAttributes.timestamp` does not fall within the time frame of the
+///    Cancun fork.
+pub fn validate_parent_beacon_block_root_presence(
+    chain_spec: &ChainSpec,
+    version: EngineApiMessageVersion,
+    timestamp: u64,
+    has_parent_beacon_block_root: bool,
+) -> Result<(), AttributesValidationError> {
+    // 1. Client software **MUST** check that provided set of parameters and their fields strictly
+    //    matches the expected one and return `-32602: Invalid params` error if this check fails.
+    //    Any field having `null` value **MUST** be considered as not provided.
+    match version {
+        EngineApiMessageVersion::V1 | EngineApiMessageVersion::V2 => {
+            if has_parent_beacon_block_root {
+                return Err(AttributesValidationError::ParentBeaconBlockRootNotSupportedBeforeV3)
+            }
+        }
+        EngineApiMessageVersion::V3 => {
+            if !has_parent_beacon_block_root {
+                return Err(AttributesValidationError::NoParentBeaconBlockRootPostCancun)
+            }
+        }
+    };
+
+    // 2. Client software **MUST** return `-38005: Unsupported fork` error if the
+    //    `payloadAttributes` is set and the `payloadAttributes.timestamp` does not fall within the
+    //    time frame of the Cancun fork.
+    validate_payload_timestamp(chain_spec, version, timestamp)?;
+
+    Ok(())
+}
+
+/// Validates the presence or exclusion of fork-specific fields based on the payload attributes
+/// and the message version.
+pub fn validate_version_specific_fields<Type>(
+    chain_spec: &ChainSpec,
+    version: EngineApiMessageVersion,
+    payload_or_attrs: &PayloadOrAttributes<'_, Type>,
+) -> Result<(), AttributesValidationError>
+where
+    Type: PayloadAttributesTrait,
+{
+    validate_withdrawals_presence(
+        chain_spec,
+        version,
+        payload_or_attrs.timestamp(),
+        payload_or_attrs.withdrawals().is_some(),
+    )?;
+    validate_parent_beacon_block_root_presence(
+        chain_spec,
+        version,
+        payload_or_attrs.timestamp(),
+        payload_or_attrs.parent_beacon_block_root().is_some(),
+    )
+}
+
+/// The version of Engine API message.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum EngineApiMessageVersion {
+    /// Version 1
+    V1,
+    /// Version 2
+    ///
+    /// Added for shanghai hardfork.
+    V2,
+    /// Version 3
+    ///
+    /// Added for cancun hardfork.
+    V3,
 }
