+++ crates/node-api/src/lib.rs
@@ -7,9 +7,13 @@
 )]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
-/// Contains the [EngineTypes] trait, used to configure the types used by the engine.
+/// Contains the [EngineTypes] trait, used to configure the types used by the engine. Also contains
+/// methods that can be used to validate payload attributes.
 pub mod engine;
-pub use engine::EngineTypes;
+pub use engine::{
+    validate_payload_timestamp, validate_version_specific_fields, validate_withdrawals_presence,
+    EngineApiMessageVersion, EngineTypes,
+};
 
 /// Contains traits and types used to abstract over payload attributes types.
 pub mod payload_attributes;
@@ -18,14 +22,6 @@ pub use payload_attributes::{PayloadAttributesTrait, PayloadBuilderAttributesTra
 /// Contains error types used in the traits defined in this crate.
 pub mod error;
 pub use error::AttributesValidationError;
-
-/// Contains types and methods that are used to validate types that implement
-/// [PayloadAttributesTrait].
-pub mod attributes_validation;
-pub use attributes_validation::{
-    validate_payload_timestamp, validate_version_specific_fields, validate_withdrawals_presence,
-    EngineApiMessageVersion,
-};
 
 /// Contains types used in implementations of [PayloadAttributesTrait].
 pub mod payload;
